Change filter for prediction
(*NOTE* This bug was found on windows platform!)

Always use "FilteredClassifier" during prediction instead of the
a dedicated one for the classifier. FilteredClassifier is used
in every configurations and therefor stored in the saved model file
(along with the whole filter configuration) as well. Specifying a different
one for prediction is recognized by weka and rejected.